Extract bookmarks and threads fetch request creation from view controllers.
